Refactored select all logic
Select all button and the list of checkboxes had a bug where it took a few clicks for it to work and then it was fine.

This has been partially fixed. Now it does what it is supposed to for the most part and the only problem is a visual glitch which may be indicative of a fault within the logic.

Need to comeback to this after implementing search bar directive.