If applied, this commit will let map display properly.
Adding search bar in bbc48bf41e5498f1291748651c2405bf9bbf4563 broke the search module as the controller for the states in the search module had the same name as the controller in the search-bar directive.

Since the search-bar directive was loaded after the search module's files, the controller from the search module was overwritten.

Changing the search-bar directive's controller's name fixed this problem.