Markers show up properly
Controller was trying to access a property that did not exist due to the variable referring to the wrapper object around the data that it was actually trying to get at. Pointing it at the right object fixed the problem.

The API service also had the wrong endpoint when trying to get sightings which is why no markers were showing up.