Completes Auth0 authentication implementation
The audience that we wanted the request to go, was not being passed into the angularAuth0Provider's init function in app.js, despite it being available in the client_config.js file. Due to this the SPWA was getting a truncated access token.

Passing in the audience parameter helped configure Auth0 in the SPWA properly.

The authentication now works properly as the access token is not truncated and also contains the scopes that were requested.