Improves UI
A button that takes the user to the admin page and a button that lets the user log out, show up on the nav bar after the user has logged in and authenticated. They also disappear when the user logs out.

The base element in the head of index.html replaces the url when deployed, with the future name of the repository (will be changed after this commit).

The select all checkbox on the admin delete pages is deselected after a delete.