Testing fix for deployment callback page not found error
A div wraps the app as the script that handles the redirect needs to be in the index.html, outside (adjacent) of the element that has the ng-app directive.

404.html redirects to the callback state properly.

Search bar directive module no longer declares a controller. Instead it refers to the controller function that is in the same file.