Adds callback state for testing deployed app redirects
Adds a callback state in the app to redirect to once the local login is done after the accessToken hash has been parsed.

This callback state will use AngularJS' $state service to redirect to the admin state which is the state that the SPWA is supposed to "unlock" after the user has authenticated.