Authenticated supported SPWA
This is the final version of the SPWA which gets an access token, with admin scope, from Auth0, and sends it to the authenticated API, with every request that it makes to the API.

The client variables have also been adjusted to ask the same Auth0 tenant, that the deployed authenticated API is registered to, for an access token. This way the API will not reject the request because it was authenticated by a different tenant.